Move generateRandomISBN13 to separate hepler file
